fixing -Wall -Wextra -Werror

+++ Parsing/parsing_help4.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   parsing_help4.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sacharai <sacharai@student.42.fr>          +#+  +:+       +#+        */
+/*   By: ojebbari <ojebbari@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/03/26 22:05:45 by sacharai          #+#    #+#             */
-/*   Updated: 2024/03/27 07:08:37 by sacharai         ###   ########.fr       */
+/*   Updated: 2024/03/27 14:03:41 by ojebbari         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ char	*create_newmap_row(char **map, int max_size_line, int row_index)
 	k = 0;
 	while (k < max_size_line + 2)
 	{
-		if (k == 0 || k > ft_strlen(map[row_index - 1]))
+		if (k == 0 || (size_t)k > ft_strlen(map[row_index - 1]))
 			newmap_row[k] = ' ';
 		else
 			newmap_row[k] = map[row_index - 1][k - 1];
